Bash: prevent `npm` output when setting $PATH
If npm was configured with a high loglevel (as is now the default in
homebrew's install of node), running `npm root` may print additional
debugging info to STDERR. Since we call `npm root` to set the $PATH to
include local npm bin directories, and we do that every prompt display
(via $PROMPT_COMMAND), this could print a ton of junk in your terminal
every time the prompt is displayed.

This commit redirects `npm`'s STDERR to `/dev/null` to prevent this.