Bash: only config prompt if `$TERM != dumb`
This is in response to `tput` not reacting well when run in a script
called from SourceTree. The $TERM in that case is `dumb` and termcap
doesn't have a database capable of a lot of the stuff we want to do
(colors, etc.) in that case.

This commit adds a `if...fi` block around a lot of interactive mode
setup commands in ~/.bash_profile, and indents the code inside by
another level. For this reason, a lot of lines are marked as 'changed'
in this commit only because their indentation changed. Be warned.