Remove git email filter
This wasn't really being used (the currently configured email in git
is, in fact, my public GitHub email, and the only other use of my
private email was in ~/.hgrc. I changed that file to use my public
email, and now the email filter is redundant.

It also wasn't really implemented right, so that's another reason to
toss it.