Git: create common gitconfig and include system-specific config
Previously, ~/.gitsupport/gitconfig.{mac,linux} were independent,
self-contained files that had all the git configuration to run on Mac
or Linux. This meant there was a lot of duplicate configuration between
the two. We then manually symlinked one or the other to ~/.gitconfig.

We now add a ~/.gitconfig file to the repo containing all the common
configuration. It has an include directive for
~/.gitsupport/gitconfig.active, which should be manually symlinked to
~/.gitsupport/gitconfig.mac or ~/.gitsupport/gitconfig.linux. These two
OS-specific files contain just the configuration specific to that OS
(for example, the Mac one has configuration for SourceTree.)