Bash: add-and-comment-out input mode indicator (conflicts with iTerm2)
This normally puts a ":" or "+" at the start of your prompt based on if
you're in Vim command mode, or Vim insertion mode (or "@" if you're in
Emacs mode, for some strange reason.)

However, this conflicts with iTerm2's shell integration, causing a
newline and prompt re-render every time the mode indicator changes. So
until that is resolved, comment the setting out, but leave it in our
`~/.inputrc` file as a reference for later.