Bash: remove `ssh` alias setting term title; ssh LocalCommand does it.
We previously had a wrapper function for `ssh` (which we aliased the
`ssh` command to) which set the Terminal title to `user@host`. However,
instead of using a janky wrapper which tries to parse the SSH args
itself, we now use a `LocalCommand` directive in ~/.ssh/config` which
will execute a command to change the terminal title once it connects.