Git: don't hard-wrap commit messages at 72 lines
I know this makes it a lot easier to read commit messages in a
terminal, and is the Linux/Git standard style for commit messages...
BUT! It makes it look really ugly in GUIs like GitHub and SourceTree,
which is the main way I actually interact with Git these days. (We
worked around the SourceTree ugliness by defining a SourceTree commit
message transform command to unwrap the messages, but that's a hack.)

We now disable this 'feature' and leave the commit messages unmodified
from what is typed.