Update brew-autoupdate to try to fix bug where agent would fail to run
There were non-zero exit codes in the launchd logs for the autoupdate
agent job. Casual observation was also showing that we were never
seeing update notifications, even when there were available updates
(meaning the job probably wasn't running and/or completing.)

This update changes some of the options in the launchd.plist config
file that runs the agent. Mainly, we delete options telling launchd to
run the process with low priority, and set up the environmental
variables to match those present when the agent was installed.