Bash: improve how we display Ubuntu dynamic MOTD
Before we were just `cat`ing `/var/run/motd.dynamic`, would could be
stale. Now we directly use `run-parts` to run the scripts in
`/etc/update-motd.d` to show the MOTD.

The only downside of this is that, if some scripts require root access
(which they usually have, because they're run by pam_motd), they will
fail. We `2>/dev/null` the command to filter out the error, but we also
don't see the intended output. So it's a trade off...