`~/.janus/update.sh`: Run submodule update commands in parallel
We loop through every submodule in the YCM directory in order to pull
its latest changes and update its own submodules. Previously, we used
only the `git submodule foreach` command to do this, which ran the
update of each submodule in serial. We now only use git to print the
submodules' paths, then use `xargs` in parallel mode (`-P n`) to run
the pull/update commands. We run `n` commands in parallel at a time,
where `n` is equal to the number of CPUs on the current system.