In `~/.janus/update.sh`, build YCM if it's binary doesn't exist at all
Previously, we added logic to `update.sh` to only (re)build YCM if the
checked out commit of its submodule changed. This meant we didn't
unnecessarily rebuild it if it was installed and we ran `update.sh`,
but YCM wasn't changed. However, it didn't account for the case of YCM
never being built in the first place (e.g., in the case of a fresh
install). In this case, unless the YCM submodule happened to be updated
in the repo, we wouldn't build YCM, despite the fact that it was never
built at all on the system.

This commit adds a check for the YCM binary file (or one of them at
least,) and, if it doesn't exist, runs a build.