Add the user Ruby Gem bin directory to $PATH
This fixes a problem where the `castle` alias wasn't working. We check that the `homesick` command is available before setting that alias. If homesick was installed by the user to the user's personal Ruby Gem directory in their home folder, it won't be in the $PATH, won't be found when the alias is to be set, and the alias won't be set.

This commit allows the `homesick` command to work as normal, even if installed to the user's Ruby Gem directory, and for the `castle` alias to work as normal.