Update `ps` alias to find $TERM width on run
Our `ps` alias explicitly sets the width of `ps`'s output, since piping
through `grc` prevents `ps` from detecting the terminal width
automatically. Instead, our alias uses `tput` to find the $TERM width,
and passes that into the call to `ps` explicitly.

Previously, some unescaped quotes meant that we were finding the $TERM
width at login (when the `ps` alias is defined) and then hard-coding
that width into the alias. What we really intended to do was find the
current width of $TERM every time we call `ps`, and use that.

This login-time $TERM width query (via `tput cols`) was also causing
warnings with `scp`, since `scp` doesn't create a $TERM, and therefore
the call to `tput cols` would print a warning when it was called during
login.

This commit properly escapes the quotes in the alias definition so that
we call `tput cols` when the `ps` command is run, not when the alias is
defined. This fixes the warning with `scp`, since now `tput` is only
called if/when `ps` is called, not during login.