Update .tmux.conf to use $HOME not hard-coded path
Previously, our `~/.tmux.conf` sourced the Powerline init script (which
resides in `~/.local/...`) using a hard-coded path
(`/home/DSAC/g.matt/...`). This meant that if my home directory was at
a different location on some other system, that command would fail, and
no Powerline goodness in tmux for me :(

This updates that `source` command in `.tmux.conf` to use the `$HOME`
environmental variable for my home directory, and quotes the path with
double-quotes so that variable expansion happens.

Now Powerline + tmux work on all my systems!