Tell Git to ignore submodule working tree changes
In this module's `.gitmodules` file, we set `ignore = dirty` for every submodule (which are all Vim plugins, in this repo's case). This tells Git to ignore any changes made to the checked out files of the submodule, so that `git diff` stops being annoying if a Vim plugin creates build products, temp files, etc.

Caveat: submodules are more complicated than they should be, so my understanding of what `ignore = dirty` does may be wrong.