Git config: configure "sourcetree" difftool/mergetool for Araxis Merge
SourceTree automatically adds `sourcetree` difftool and mergetool entries to `~/.gitconfig`. The `cmd` used for both of these entries reflects the external diff/merge program you have configured in the SourceTree GUI. We currently have SourceTree configured to use Araxis Merge as it's external diff/merge program, and so this commit updates `.gitconfig` for that setting.

Mainly, this commit is here so that `~/.gitconfig` isn't eternally marked as "changed" :P