In Atom's `styles.less`, fix `@import` bug caused by Homesick tracking
In `styles.less`, we import some `file-icons` package files. Normally, we could use the relative directory of `packages/file-icons/...` for this import (which is what we had previously), since the import was done relative to the `styles.less` file, which was in the `~/.atom/` directory.

However, now that Homesick is tracking the `styles.less` file, it has been moved to `~/.homesick/repos/dotfiles/home/.atom` (with a symlink from the old location to the Homesick location). This broke the `@import` statement in `styles.less`.

This commit fixes the import path to traverse to the '~' directory, using relative `..` parent path components, then refer to the files in `.atom/packages/file-icons/`. This isn't pretty looking, but it should work no matter the path to our home directory.