Jupyter: remove custom 'open browser' command from config
Previously, we had to hack the 'open browser' command Jupyter Notebook would run on startup, so that it opened to the default URL on startup.

The newest version of Notebook seems to open the browser to the default URL automatically, so we don't need the previous hack. In fact, with this new behavior, explicitly appending the default URL to the 'open browser' command causes a 404 error when the browser opens, since it will try to load 'localhost/<default URL>/<default URL>', which isn't a valid address.