bin/helpers.sh: `check_url` now accounts for a bug with Apple's site
Previously, using the `check_url` function in `~/bin/helpers.sh` to check a URL on https://developer.apple.com (as the `hman` command does) could erroneously return success, even if opening the URL in the browser would show a 404 error page. This is caused by Apple (for some stupid reason) returning a 300 HTTP response code, when it meant to show a 404.

We now check the HTTP response code (in addition to `curl`'s exit code) to see if it's `300`. If it is, we return a non-zero value from the function, indicating failure.