scat: support multiple files, render Markdown in terminal
`scat` has had a large refactoring. The first new feature is being able to pass multiple input files to the command, and having it handle that gracefully (displaying one after the other).

The second new feature is rendering Markdown files in the terminal, by first using `pandoc` to convert the file to a man page, then using `groff` to display the man page in the terminal.

We also have a new method for displaying files that fit on one page of the terminal. Previously, we manually ran `wc -l` on the input file and compared that to the current height (in rows) of the terminal. If the file fit on one page, we displayed it with `cat`; otherwise, we used `less`. Now, we always use `less`, but pass the `-XF` arguments to it, which tells it to exit immediately if the input fits in the terminal, and to run as normal otherwise.