Atom - styles: set `@ui-theme-name` var to default, in case not set
Most themes set the LESS variable `@ui-theme-name` to be their name, and so we can use that variable in conditionals to check if the theme is active. However, some don't set this variable, causing an error when we try to access it (even if the conditional would otherwise be false).

This commit sets that variable to "unknown" before importing the UI theme variables of the current theme. That way, if the theme doesn't set it, accessing it later doesn't cause an error.