Add Brewfile, for automated install of my Homebrew formulas (WIP)
This is some text that does some stuff and like totally is awesome so you

This commit adds a file named `Brewfile` to the root of this repo. This file can be used with by Brew Bundle (https://github.com/Homebrew/homebrew-bundle) to automatically execute Homebrew commands.

Specifically, this is useful for automatically installing all my favorite/required software on a new machine. It can be seen as the dual of the dotfiles in this repo; `homesick` uses this repo to automatically configure the settings of a new machine, and `brew bundle` is used to automatically install the software of a new machine. Together, they should be able to automatically setup a fresh Mac OS X machine exactly how I want, so I can immediately begin working.

To use this Brewfile, you need to first install Homebrew Bundle: `brew tap Homebrew/bundle`. Then run `brew bundle ~/.homesick/repos/dotfiles/Brewfile`.

The version of the file in this commit should be considered a WIP. It was generated by starting with the output of `brew bundle dump` (which creates a Brewfile based off the list of currently formulas/taps), then manually cherry-picking the formulas I cared about from the noise of auto-installed dependencies and such. However, I haven't actually run the Brewfile on a fresh OS X install, so I can't verify it works or is complete. So caveat emptor.

Possible improvements to the Brewfile could include leveraging Homebrew Casks and/or Brew Bundle's Mac App Store integration. This would allow me to automatically install *all* my software on a new machine, including Mac OS X .app's and other GUI software, instead of being limited to the CLI-focused formulas in the standard Homebrew taps. Additionally, I could

This file won't be symlinked anywhere by Homesick (since it's not in the `./home/` subdirectory), but can be accessed in `~/.homesick/repos/dotfiles/Brewfile`.