Git: have MacVim run Terminal.app, not iTerm, after editing commit msg
The Mac-specific Git config file has a commented-out line configuring the commit message editor to be MacVim. It's currently commented-out in favor of Atom as the editor, but it's still there for reference, if we want to switch back.

This commit changes that commented-out config so that MacVim is run with arguments telling it to open Terminal.app when done, not iTerm. This argument is useful because, when you're done editing the commit message and quit MacVim, you're automatically sent back to the Terminal to see Git complete the commit. Since we switched our default terminal emulator back to Terminal.app, from iTerm, months ago, we change this command.