Atom: in "increment font sizes", don't set UI fontsize if not one-*-UI
In our atom `init.coffee` file, we setup a command (which our keymap maps a shortcut to) to increment both the editor font size *and* UI font size (by default, Atom just increases the editor font size). However, only the `one-light-ui` and `one-dark-ui` UI themes have a `fontSize` setting we can set; setting the option for other UI themes has no effect, other than to create a meaningless config setting.

This commit updates the code for that command (in `init.coffee`) so that it checks if the current UI theme is `one-*`, and only sets the UI `fontSize` setting if so. If another UI theme is detected, the command returns without doing anything.