Atom: make cursor location status bar widget less likely to change size\
The "cursor location" widget in Atom's status bar contains text like "Line 36, Column 23". As a side effect, as you move the cursor around (especially changing the digits in the line/column number, like "Line 50" to "Line 2"), the widget changes size (to contain the current text). This, in turn, caused all the widgets in the status bar to shift position left/right as you moved around, which was distracting.

This commit gives that widget a min-width large enough to hold 3 digits for both line and column position. So long as you stay below line/column 1000, the widget shouldn't change size. This cuts down on visual distraction.