Atom: in "git commit message" files, don't make "overly long" lines red
The syntax definition for the "git commit message" file type (which is used when Atom is set as Git's commit message editor) parses any characters in a line >=70 as "invalid"/"illegal". This causes syntax themes to give that text angry styling (bold black on red, for example). The purpose of this is to force you to hard-wrap your lines in commit messages to be short.

However, for reasons described in the code comments of this commit, this is stupid. Atom can't automatically hard-wrap lines, and doing that manually sucks. Further, hard-wrapping commit messages is silly, since Git UIs displaying commit logs should soft-wrap messages on display. This allows them to wrap the message at the width of the current UI, not at a pre-determined 70 characters. Since we use SourceTreee as our Git UI, and its commit message width varies with the size of the window, this matters to us.

This commit simply undoes the styling of "invalid"/"illegal" tokens in git commit message files. Since the syntax definition is built into Atom, it's hard to edit it so that it will parse the lines correctly. Therefore, the easiest solution is to change the styling of text so that those tokens have no special styling, making them appear as regular text.