Bash: on Mac, disable `nvm`, so that we can instead use `n`
I think I like `n` better than `nvm` for a Node.js version manager. So, on Mac at least, disable `nvm`'s initialization in our `.bash_profile`, so that it won't be used. The `node` command will, therefore, default to the version of `node` found in the `$PATH`. I use the `n` utility to install/set the version of `node` found in `/usr/local`, so the version in the `$PATH` will be the version managed by `n`.