Jupyter Notebook: use ~/.jupyter for CWD, unless $PWD has node_modules
In our Jupyter Notebook config file, we check the current directory to see if it has a `node_modules` directory in it. If so, we use the current directory as a working directory in Jupyter. Otherwise, we use `~/.jupyter`.

This is useful when using the Node.js kernel in Jupyter Notebook. We can install npm modules to the local directory, then run Jupyter and `require()` them easily. But if we don't have local modules, we use a common location as the working directory, allowing us to install modules globally, that we can always use from Jupyter.