Atom: fix three-finger swipe gestures not working
There is a bug in Atom where, if a user is actively scrolling with the mouse wheel, programmatic movement commands (like scroll to top/bottom of a file) will be ignored (though the cursor position will still be moved).

Because of this bug, my BetterTouchTool config mapping "three finger swipe down/up" to "end"/"home" doesn't work in Atom; BTT doesn't register the gesture immediately, so Atom sees a scroll event start, then BTT registers the gesture and sends "end"/"home", but Atom ignores this because a scroll is already being handled.

It's really hard to pin down the exact cause of the bug, but I *think* it's because the "end" keypress event is mapped into 'move cursor to end, then autoscroll editor to cursor'. Autoscroll and 'regular' scroll seem to be handled differently in `text-editor-presenter.coffee`, and calling the function sets a "desired auto/regular scroll position" variable, clears the "desired scroll position" variable for the other type of scroll, then emits an event to tell the editor to update itself. It seems like "end" sets the desired autoscroll position, but before the editor can update itself, "mouse wheel scroll" sets its own desired position and clears the autoscroll. Then, when the editor update finally happens, it only sees the "mouse wheel scroll" position update.

My hacky attempt at a workaround is contained in this commit. In short, as soon as we send the "move to top" command, we tell the editor to update its position, before any mouse wheel scroll events can be processed and mess with the "end"/"home" autoscroll.