Atom: refactor custom styling of minimaps, to fix subtle bugs
This commit refactors the custom styling `styles.less` applies to minimap elements (changing background color, visible area color/borders, and adding a border on the left of the minimap). These new methods of styling the element should eliminate some subtle bugs we were seeing before, with the element being a few pixels out of position, or acting funny if toggled on/off/on.