Atom: in init.coffee, add code for advanced management of antialiasing
On non-Retina screens, I want to use subpixel antialiasing of text in Atom's text editors. However, subpixel AA looks bad when used with a bold font.

To solve this, I created a new class (in init.coffee), of which a new instance is constructed on startup. Once initialized, this class will examine all the styles set by the syntax theme, and find any that set `font-weight: bold`. It gets the selectors of all such rules, and combines them to create a selector that selects all scopes with a bold font. Finally, it adds a new, programmatically-generated style sheet to Atom which contains this 'bold' selector, and sets `-webkit-font-smoothing: antialias !important` (i.e., disables subpixel AA, and uses regular AA).

The class also watches for syntax theme changes, and re-runs itself to update its stylesheet.

This functionality should probably be broken out into its own package, as it's straining the limits of reasonableness for init.coffee's intended feature scope and size. For now, though, it lives heres.