Atom: make highlight of current line more subtle
We already use the `highlight-line` package to add a highlight on whatever line the cursor is located on. (Some syntax themes do this automatically, but it's variable enough to be worthwhile installing this package to ensure that we always get that highlighting, regardless of the current theme.)

However, the default background color used to highlight the line is pretty contrast-y. This becomes a big problem when the syntax theme makes some text dark-ish on a dark background (or light-ish on a light background). In this case, the highlight-line background color makes that text incredibly difficult to read on the current, highlighted line. This is a common technique a lot of themes use to make things like comments less distracting from code.

This commit adds a style rule to our user stylesheet which sets the background color of the highlighted line to be subtler. Specifically, it uses the background color set by the theme for highlighting line numbers of "selected" lines. This is usually chosen by the theme author to be a subtle but obvious color, so it works well for highlight-line.