Atom: remove custom .js file icon; package now uses that icon standard
Previously, the default icon assigned by the `file-icons` package to `.js` (and related) files was bad; it was drawn as a thin outline of shapes, that was hard to read and indistinct at the sizes it was used for.

To make this better, I switched the icon (using stylesheet rules in my styles.less) to be the same basic shape, but use a filled-in box (rather than an outlined box). This was much easier to read/see. I also submitted a PR to the file-icons package to make this the default.

Now, months later, file-icons finally accepted my PR (well, they actually closed my PR and made equivalent changes themselves), and the ‘filled-in box’ JS icon is now the default. Therefore, I don’t need these custom tweaks in my styles.less anymore.