`hman`: don't load system man pages from developer.apple.com
Previously, `hman` took advantage of the fact that Apple's developer website (https://developer.apple.com) already had HTML versions of man pages for built-in OS X software. If it detected that the man page being requested was part of the OS, it would attempt to find its man page on Apple's site and open that, rather than manually convert the local man page to HTML and open that.

However, a few years ago, Apple marked all the man pages on their site as "legacy" documentation, and stopped updating. At this point, the online man pages are significantly out-of-date compared to the software included in macOS Sierra. In many instances, the online man page has incorrect information, or is at least missing options/features, when using the current version of software.

The local man pages on-disk continue to be updated to match the software included with Sierra. Therefore, they should be considered the only source of 'truth' for this software, and the online man pages should no longer be used.

This commit removes the code from `hman` that detected if the man page being requested was part of the standard OS X installation, then tried to find and open its corresponding man page on https://developer.apple.com. Sans this code, the default behavior of manually converting the local man page to HTML will be used, as it already does with non-standard (e.g. Homebrew-installed) software.

This is a shame, because our local 'man page -> HTML' converter kinda sucks. Apple's online man pages were at least prettier and easier to read, if a little bare-bones. This might be good incentive to improve our converter, but man pages are pretty awful to deal with and lack virtually any semantic markup, so there's a limit on what can be done.

`hman` continues to open Git man pages using the web-based version on http://git-scm.com, if available.