Vim: in `~/.janus/update.sh`, do minimal work if `--no-upgrades` given
This commit wraps most of the work of updating Vim plugin submodules in an `if` statement.

If the `--no-upgrades` option is *not* passed to the script, then it will do the full suite of work it previously did: quit Atom, stash changes, upgrade submodules, commit changes, and unstash changes.

If the `--no-upgrades` option *is* passed to the script, then it will only fetch and checkout each submodule to the revision currently specified in the current revision of this (dotfiles) repo.