Disable notifications in `brew autoupdate`
Our `brew autoupdate` script relied on the program `terminal-notifier`
to display notifications in macOS' Notification Center.

However, `terminal-notifier` has been fatally broken in recent versions,
and just hangs forever when run, never doing anything, never returning.
This caused our script to hang forever, waiting for `terminal-notifier`
to exit.

To temporarily fix this, I've removed the call to the
`terminal-notifier` command. This means the script can no longer show
notifications. This, arguably, makes it useless. But oh well, maybe it
can be fixed in the future...

This commit also adds error handling to the `brew autoupdate uninstall`
subcommand, which checks to make sure the agent is actually installed
before trying to uninstall it. If it isn't, it just shows an error
message and exits.